backend/requirements.txt
fastapi
uvicorn
pydantic
Pillow
sentence-transformers
chromadb
deep-translator
python-multipart